requirements.txt
matplotlib==3.3.4
numpy==1.18.4
pandas==0.23.0
Pillow==11.0.0
PyMySQL==1.0.2
PyYAML==6.0.2
SQLAlchemy==1.2.7
